requirements.txt
pylint
pytest
codecov
pytest-cov
coverage
ruff
pytest-mock
fastapi
uvicorn[standard]
sqlalchemy
databases
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
bcrypt
pydantic[email]
psycopg2-binary
asyncpg
python-dotenv